Abstract ReadDescription
- Remove ReadDescription logic from index.js
- Remove GET_DESCRIPTION messages from index.js
- Export messages and handlers in getDescriptionIntent.js
- Register getDescription handler for Alexa
